feat: add RouterSelect component for improved router management and update App component to integrate it
